QF-1001 Change unit tests
